requirement.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
typing-extensions==4.5.0
